refactor: split rss/feed parsing and settings/custom_formats
- Extract rss.rs feed fetch + XML parse into services/rss/feed.rs (275 lines).
  mod.rs shrinks from 1561 to 1306 lines.
- Extract settings.rs Custom Format CRUD/import/export handlers + helpers
  into handlers/settings/custom_formats.rs (1020 lines). mod.rs shrinks from
  2400 to 872 lines. ImportCollision/ImportPreviewEntry/ImportReviewView move
  with the CF code; SettingsTemplate gets pub(super) visibility so the import
  handler can still re-render the review inline via build_settings_template.
- main.rs OpenApi references bump to handlers::settings::custom_formats::*
  because utoipa's __path_X types live at the module where the #[utoipa::path]
  attribute applies, not where the function is re-exported.

No behavior change. cargo test still 480 passes.
