+++ src/main.rs
@@ -79,14 +79,14 @@ use services::{
         handlers::settings::jellyfin_test,
         handlers::settings::jellyfin_refresh,
         // Settings — Custom Formats
-        handlers::settings::settings_custom_formats_upsert,
-        handlers::settings::settings_custom_formats_delete,
-        handlers::settings::settings_custom_formats_minimum_score,
-        handlers::settings::settings_custom_formats_import,
-        handlers::settings::settings_custom_formats_import_resolve,
-        handlers::settings::settings_custom_formats_install_defaults,
-        handlers::settings::settings_custom_formats_reset_defaults,
-        handlers::settings::settings_custom_formats_export,
+        handlers::settings::custom_formats::settings_custom_formats_upsert,
+        handlers::settings::custom_formats::settings_custom_formats_delete,
+        handlers::settings::custom_formats::settings_custom_formats_minimum_score,
+        handlers::settings::custom_formats::settings_custom_formats_import,
+        handlers::settings::custom_formats::settings_custom_formats_import_resolve,
+        handlers::settings::custom_formats::settings_custom_formats_install_defaults,
+        handlers::settings::custom_formats::settings_custom_formats_reset_defaults,
+        handlers::settings::custom_formats::settings_custom_formats_export,
         handlers::system::api_logs_poll,
         handlers::system::api_logs_clear,
         handlers::system::api_logs_client,
